refactor: editor indent functions to IndentHelper
